backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
fastapi-cors==0.0.6
sqlalchemy==2.0.35
python-dotenv==1.0.0
google-generativeai==0.7.2
requests==2.31.0
